backend/requirements.txt
# ============================================================
# AgentNews Backend - Python Dependencies
# ============================================================

# ---- Web Framework ----
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ---- Database ----
aiosqlite>=0.19.0
# motor>=3.6.0  <-- MongoDB removed for portability
# pymongo>=4.9.0

# ---- Scheduling ----
apscheduler>=3.10.0

# ---- Data Ingestion ----
feedparser>=6.0.0
httpx>=0.27.0
markdownify>=0.13.0
beautifulsoup4>=4.12.0

# ---- LLM Providers ----
openai>=1.50.0
anthropic>=0.37.0

# ---- Data Validation ----
pydantic>=2.9.0

# ---- Environment ----
python-dotenv>=1.0.0

# ---- Optional: JS-rendered scraping ----
# playwright>=1.48.0  (install separately: pip install playwright && playwright install)
